+++ src/room/participant/LocalParticipant.ts
@@ -17,10 +17,12 @@ import LocalTrack from '../track/LocalTrack';
 import LocalTrackPublication from '../track/LocalTrackPublication';
 import LocalVideoTrack, { videoLayersFromEncodings } from '../track/LocalVideoTrack';
 import {
+  AudioCaptureOptions,
   CreateLocalTracksOptions,
   ScreenShareCaptureOptions,
   ScreenSharePresets,
   TrackPublishOptions,
+  VideoCaptureOptions,
   VideoCodec,
 } from '../track/options';
 import { Track } from '../track/Track';
@@ -31,7 +33,7 @@ import { ParticipantTrackPermission, trackPermissionToProto } from './Participan
 import { computeVideoEncodings, mediaTrackToLocalTrack } from './publishUtils';
 import RemoteParticipant from './RemoteParticipant';
 
-const compatibleCodecForSVC = 'vp8';
+const compatibleCodec = 'vp8';
 export default class LocalParticipant extends Participant {
   audioTracks: Map<string, LocalTrackPublication>;
 
@@ -114,8 +116,11 @@ export default class LocalParticipant extends Participant {
    * If a track has already published, it'll mute or unmute the track.
    * Resolves with a `LocalTrackPublication` instance if successful and `undefined` otherwise
    */
-  setCameraEnabled(enabled: boolean): Promise<LocalTrackPublication | undefined> {
-    return this.setTrackEnabled(Track.Source.Camera, enabled);
+  setCameraEnabled(
+    enabled: boolean,
+    options?: VideoCaptureOptions,
+  ): Promise<LocalTrackPublication | undefined> {
+    return this.setTrackEnabled(Track.Source.Camera, enabled, options);
   }
 
   /**
@@ -124,16 +129,22 @@ export default class LocalParticipant extends Participant {
    * If a track has already published, it'll mute or unmute the track.
    * Resolves with a `LocalTrackPublication` instance if successful and `undefined` otherwise
    */
-  setMicrophoneEnabled(enabled: boolean): Promise<LocalTrackPublication | undefined> {
-    return this.setTrackEnabled(Track.Source.Microphone, enabled);
+  setMicrophoneEnabled(
+    enabled: boolean,
+    options?: AudioCaptureOptions,
+  ): Promise<LocalTrackPublication | undefined> {
+    return this.setTrackEnabled(Track.Source.Microphone, enabled, options);
   }
 
   /**
    * Start or stop sharing a participant's screen
    * Resolves with a `LocalTrackPublication` instance if successful and `undefined` otherwise
    */
-  setScreenShareEnabled(enabled: boolean): Promise<LocalTrackPublication | undefined> {
-    return this.setTrackEnabled(Track.Source.ScreenShare, enabled);
+  setScreenShareEnabled(
+    enabled: boolean,
+    options?: ScreenShareCaptureOptions,
+  ): Promise<LocalTrackPublication | undefined> {
+    return this.setTrackEnabled(Track.Source.ScreenShare, enabled, options);
   }
 
   /** @internal */
@@ -152,16 +163,32 @@ export default class LocalParticipant extends Participant {
    * Resolves with LocalTrackPublication if successful and void otherwise
    */
   private async setTrackEnabled(
-    source: Track.Source,
+    source: Extract<Track.Source, Track.Source.Camera>,
     enabled: boolean,
-  ): Promise<LocalTrackPublication | undefined> {
+    options?: VideoCaptureOptions,
+  ): Promise<LocalTrackPublication | undefined>;
+  private async setTrackEnabled(
+    source: Extract<Track.Source, Track.Source.Microphone>,
+    enabled: boolean,
+    options?: AudioCaptureOptions,
+  ): Promise<LocalTrackPublication | undefined>;
+  private async setTrackEnabled(
+    source: Extract<Track.Source, Track.Source.ScreenShare>,
+    enabled: boolean,
+    options?: ScreenShareCaptureOptions,
+  ): Promise<LocalTrackPublication | undefined>;
+  private async setTrackEnabled(
+    source: Track.Source,
+    enabled: true,
+    options?: VideoCaptureOptions | AudioCaptureOptions | ScreenShareCaptureOptions,
+  ) {
     log.debug('setTrackEnabled', { source, enabled });
     let track = this.getTrack(source);
     if (enabled) {
       if (track) {
         await track.unmute();
       } else {
-        let localTrack: LocalTrack | undefined;
+        let localTracks: Array<LocalTrack> | undefined;
         if (this.pendingPublishing.has(source)) {
           log.info('skipping duplicate published source', { source });
           // no-op it's already been requested
@@ -171,23 +198,32 @@ export default class LocalParticipant extends Participant {
         try {
           switch (source) {
             case Track.Source.Camera:
-              [localTrack] = await this.createTracks({
-                video: true,
+              localTracks = await this.createTracks({
+                video: (options as VideoCaptureOptions | undefined) ?? true,
               });
+
               break;
             case Track.Source.Microphone:
-              [localTrack] = await this.createTracks({
-                audio: true,
+              localTracks = await this.createTracks({
+                audio: (options as AudioCaptureOptions | undefined) ?? true,
               });
               break;
             case Track.Source.ScreenShare:
-              [localTrack] = await this.createScreenTracks({ audio: false });
+              localTracks = await this.createScreenTracks({
+                ...(options as ScreenShareCaptureOptions | undefined),
+              });
               break;
             default:
               throw new TrackInvalidError(source);
           }
-
-          track = await this.publishTrack(localTrack);
+          const publishPromises: Array<Promise<LocalTrackPublication>> = [];
+          for (const localTrack of localTracks) {
+            publishPromises.push(this.publishTrack(localTrack));
+          }
+          const publishedTracks = await Promise.all(publishPromises);
+          // for screen share publications including audio, this will only return the screen share publication, not the screen share audio one
+          // revisit if we want to return an array of tracks instead for v2
+          [track] = publishedTracks;
         } catch (e) {
           if (e instanceof Error && !(e instanceof TrackInvalidError)) {
             this.emit(ParticipantEvent.MediaDevicesError, e);
@@ -201,6 +237,10 @@ export default class LocalParticipant extends Participant {
       // screenshare cannot be muted, unpublish instead
       if (source === Track.Source.ScreenShare) {
         track = this.unpublishTrack(track.track);
+        const screenAudioTrack = this.getTrack(Track.Source.ScreenShareAudio);
+        if (screenAudioTrack && screenAudioTrack.track) {
+          this.unpublishTrack(screenAudioTrack.track);
+        }
       } else {
         await track.mute();
       }
@@ -421,35 +461,38 @@ export default class LocalParticipant extends Participant {
       req.width = width ?? 0;
       req.height = height ?? 0;
       // for svc codecs, disable simulcast and use vp8 for backup codec
-      if (
-        track instanceof LocalVideoTrack &&
-        (opts?.videoCodec === 'vp9' || opts?.videoCodec === 'av1')
-      ) {
-        // set scalabilityMode to 'L3T3' by default
-        opts.scalabilityMode = opts.scalabilityMode ?? 'L3T3';
+      if (track instanceof LocalVideoTrack) {
+        if (opts?.videoCodec === 'vp9' || opts?.videoCodec === 'av1') {
+          // set scalabilityMode to 'L3T3' by default
+          opts.scalabilityMode = opts.scalabilityMode ?? 'L3T3';
 
-        // add backup codec track
-        const simOpts = { ...opts };
-        simOpts.simulcast = true;
-        simOpts.scalabilityMode = undefined;
-        simEncodings = computeVideoEncodings(
-          track.source === Track.Source.ScreenShare,
-          width,
-          height,
-          simOpts,
-        );
-        req.simulcastCodecs = [
-          {
-            codec: opts.videoCodec,
-            cid: track.mediaStreamTrack.id,
-            enableSimulcastLayers: true,
-          },
-          {
-            codec: compatibleCodecForSVC,
-            cid: '',
-            enableSimulcastLayers: true,
-          },
-        ];
+          // add backup codec track
+          const simOpts = { ...opts };
+          simOpts.simulcast = true;
+          simOpts.scalabilityMode = undefined;
+          simEncodings = computeVideoEncodings(
+            track.source === Track.Source.ScreenShare,
+            width,
+            height,
+            simOpts,
+          );
+        }
+
+        // set vp8 codec as backup for any other codecs
+        if (opts.videoCodec && opts.videoCodec !== 'vp8') {
+          req.simulcastCodecs = [
+            {
+              codec: opts.videoCodec,
+              cid: track.mediaStreamTrack.id,
+              enableSimulcastLayers: true,
+            },
+            {
+              codec: compatibleCodec,
+              cid: '',
+              enableSimulcastLayers: true,
+            },
+          ];
+        }
       }
 
       encodings = computeVideoEncodings(
@@ -524,6 +567,8 @@ export default class LocalParticipant extends Participant {
       ...this.roomOptions?.publishDefaults,
       ...options,
     };
+    // clear scalabilityMode setting for backup codec
+    opts.scalabilityMode = undefined;
     opts.videoCodec = videoCodec;
     // is it not published? if so skip
     let existingPublication: LocalTrackPublication | undefined;